Created "cenit" javascript module
